// src/types/color.types.ts
export interface ColorTheme {
 primaryColor: string;
 secondaryColor: string;
 tertiaryColor: string;
 sellColor: string;
 buyColor: string;
 typoPrimaryColor: string;
 typoSecondaryColor: string;
 typoAccentColor: string;
 typoAccent2Color: string;
}

export const DEFAULT_COLOR_THEME: ColorTheme = {
 primaryColor: "#0b1318",
 secondaryColor: "#012021",
 tertiaryColor: "#151f28",
 sellColor: "#e53c65",
 buyColor: "#00ff90",
 typoPrimaryColor: "#c9e4e5",
 typoSecondaryColor: "#6b9598",
 typoAccentColor: "#72eaee",
 typoAccent2Color: "#8facad",
};